bug fixed

+++ app/src/main/java/com/zhaidou/fragments/OrderReturnFragment.java
@@ -244,7 +244,7 @@ public class OrderReturnFragment extends BaseFragment implements View.OnClickLis
                             JSONObject orderObj = orderArr.optJSONObject(i);
                             int id = orderObj.optInt("id");
                             String number = orderObj.optString("number");
-                            int amount = orderObj.optInt("amount");
+                            double amount = orderObj.optDouble("amount");
                             String status = orderObj.optString("status");
                             String status_ch = orderObj.optString("status_ch");
                             String created_at = orderObj.optString("created_at");
